+++ src/de/brightslearning/java/advanced/designpattern/observer/implementation1/Channel.java
@@ -1,0 +1,5 @@
+package de.brightslearning.java.advanced.designpattern.observer.implementation1;
+
+public interface Channel {
+    void update(Object o);
+}
