--- src/de/brightslearning/java/advanced/designpattern/observer/withjava/Channel.java
@@ -1,5 +1,0 @@
-package de.brightslearning.java.advanced.designpattern.observer.withjava;
-
-public interface Channel {
-    void update(Object o);
-}
